Moved tests to gasQueue
